Remove now-redundant 'show platform' tests from test_platform_info.py
