+++ tests/platform_tests/test_platform_info.py
@@ -18,10 +18,8 @@ pytestmark = [
     pytest.mark.topology('any')
 ]
 
-CMD_PLATFORM_SUMMARY = "show platform summary"
 CMD_PLATFORM_PSUSTATUS = "show platform psustatus"
 CMD_PLATFORM_FANSTATUS = "show platform fan"
-CMD_PLATFORM_TEMPER = "show platform temperature"
 
 THERMAL_CONTROL_TEST_WAIT_TIME = 65
 THERMAL_CONTROL_TEST_CHECK_INTERVAL = 5
@@ -101,50 +99,6 @@ def psu_test_setup_teardown(duthost):
             assert False, "Failed to restart sensord task after test."
     else:
         logging.info("sensord is running, pid = {}".format(sensord_pid))
-
-
-def test_show_platform_summary(duthost):
-    """
-    @summary: Check output of 'show platform summary'
-    """
-    logging.info("Check output of '%s'" % CMD_PLATFORM_SUMMARY)
-    platform_summary = duthost.command(CMD_PLATFORM_SUMMARY)
-    expected_fields = set(["Platform", "HwSKU", "ASIC"])
-    actual_fields = set()
-    for line in platform_summary["stdout_lines"]:
-        key_value = line.split(":")
-        assert len(key_value) == 2, "output format is not 'field_name: field_value'"
-        assert len(key_value[1]) > 0, "No value for field %s" % key_value[0]
-        actual_fields.add(line.split(":")[0])
-    assert actual_fields == expected_fields, \
-        "Unexpected output fields, actual=%s, expected=%s" % (str(actual_fields), str(expected_fields))
-
-
-def check_vendor_specific_psustatus(dut, psu_status_line):
-    """
-    @summary: Vendor specific psu status check
-    """
-    if dut.facts["asic_type"] in ["mellanox"]:
-        from .mellanox.check_sysfs import check_psu_sysfs
-
-        psu_line_pattern = re.compile(r"PSU\s+(\d)+\s+(OK|NOT OK|NOT PRESENT)")
-        psu_match = psu_line_pattern.match(psu_status_line)
-        psu_id = psu_match.group(1)
-        psu_status = psu_match.group(2)
-
-        check_psu_sysfs(dut, psu_id, psu_status)
-
-
-def test_show_platform_psustatus(duthost):
-    """
-    @summary: Check output of 'show platform psustatus'
-    """
-    logging.info("Check PSU status using '%s', hostname: %s" % (CMD_PLATFORM_PSUSTATUS, duthost.hostname))
-    psu_status = duthost.command(CMD_PLATFORM_PSUSTATUS)
-    psu_line_pattern = re.compile(r"PSU\s+\d+\s+(OK|NOT OK|NOT PRESENT)")
-    for line in psu_status["stdout_lines"][2:]:
-        assert psu_line_pattern.match(line), "Unexpected PSU status output"
-        check_vendor_specific_psustatus(duthost, line)
 
 
 def get_psu_num(dut):
@@ -254,94 +208,6 @@ def test_turn_on_off_psu_and_check_psustatus(duthost, psu_controller):
         assert psu_test_results[psu], "Test psu status of PSU %s failed" % psu
 
     loganalyzer.analyze(marker)
-
-
-def parse_platform_summary(raw_input_lines):
-    """
-    @summary: Helper function for parsing the output of 'show system platform'
-    @return: Returned parsed information in a dictionary
-    """
-    res = {}
-    for line in raw_input_lines:
-        fields = line.split(":")
-        if len(fields) != 2:
-            continue
-        res[fields[0].lower()] = fields[1].strip()
-    return res
-
-
-def check_show_platform_fanstatus_output(lines):
-    """
-    @summary: Check basic output of 'show platform fan'. Expect output are:
-              "Fan Not detected" or a table of fan status data with 8 columns.
-    """
-    assert len(lines) > 0, 'There must be at least one line output for show platform fans'
-    if len(lines) == 1:
-        assert lines[0].encode('utf-8').strip() == 'Fan Not detected'
-    else:
-        assert len(lines) > 2, 'There must be at least two lines output for show platform fans if any FAN is detected'
-        second_line = lines[1]
-        field_ranges = get_field_range(second_line)
-        assert len(field_ranges) == 8, 'There must be 8 columns in output of show platform fans'
-
-
-def test_show_platform_fanstatus(duthost, mocker_factory):
-    """
-    @summary: Check output of 'show platform fan'.
-    """
-    # Do basic check first
-    logging.info("Check output of '%s'" % CMD_PLATFORM_FANSTATUS)
-    cli_fan_status = duthost.command(CMD_PLATFORM_FANSTATUS)
-    lines = cli_fan_status["stdout_lines"]
-    check_show_platform_fanstatus_output(lines)
-
-    # Mock data and check
-    mocker = mocker_factory(duthost, 'FanStatusMocker')
-    if mocker is None:
-        pytest.skip("No FanStatusMocker for %s, skip rest of the testing in this case" % duthost.facts['asic_type'])
-
-    logging.info('Mock FAN status data...')
-    mocker.mock_data()
-    logging.info('Wait and check actual data with mocked FAN status data...')
-    result = check_cli_output_with_mocker(duthost, mocker, CMD_PLATFORM_FANSTATUS, THERMAL_CONTROL_TEST_WAIT_TIME, 2)
-
-    assert result, 'FAN mock data mismatch'
-
-
-def check_show_platform_temperature_output(lines):
-    """
-    @summary: Check basic output of 'show platform temperature'. Expect output are:
-              "Thermal Not detected" or a table of thermal status data with 8 columns.
-    """
-    assert len(lines) > 0, 'There must be at least one line output for show platform temperature'
-    if len(lines) == 1:
-        assert lines[0].encode('utf-8').strip() == 'Thermal Not detected'
-    else:
-        assert len(lines) > 2, 'There must be at least two lines output for show platform temperature if any thermal is detected'
-        second_line = lines[1]
-        field_ranges = get_field_range(second_line)
-        assert len(field_ranges) == 8, 'There must be 8 columns in output of show platform temperature'
-
-
-def test_show_platform_temperature(duthost, mocker_factory):
-    """
-    @summary: Check output of 'show platform temperature'
-    """
-    # Do basic check first
-    logging.info("Check output of '%s'" % CMD_PLATFORM_TEMPER)
-    cli_thermal_status = duthost.command(CMD_PLATFORM_TEMPER)
-
-    # Mock data and check
-    mocker = mocker_factory(duthost, 'ThermalStatusMocker')
-    if mocker is None:
-        pytest.skip("No ThermalStatusMocker for %s, skip rest of the testing in this case" % duthost.facts['asic_type'])
-
-    logging.info('Mock Thermal status data...')
-    mocker.mock_data()
-    logging.info('Wait and check actual data with mocked Thermal status data...')
-    result = check_cli_output_with_mocker(duthost, mocker, CMD_PLATFORM_TEMPER, THERMAL_CONTROL_TEST_WAIT_TIME)
-
-    assert result, 'Thermal mock data mismatch'
 
 
 @pytest.mark.disable_loganalyzer
